fastapi-app/requirements.txt
fastapi
uvicorn[standard]
pytest
pytest-html
httpx
pytest-cov
requests
python-dotenv
prometheus-fastapi-instrumentator
prometheus-client
python-multipart
python-logging-loki==0.3.1